% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximatePoints.R
\name{approximatePoint}
\alias{approximatePoint}
\title{Helper to compute the Nadir point.}
\usage{
approximatePoint(..., sets = NULL, FUN)
}
\arguments{
\item{...}{[any]\cr
Arbitrary number of matrizes.}

\item{sets}{[\code{list}]\cr
List of sets. This is an alternative way of passing the sets. Can be used
exclusively or combined with \code{...}.}

\item{FUN}{[\code{function}]\cr
Either min or max function.}
}
\value{
[numeric]
}
\description{
The nadir point is constructed by the worst objective values of the
concatenation of all given point sets.
}

