% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/approximatePoints.R
\name{assertListOfPointSets}
\alias{assertListOfPointSets}
\title{Checks for point set.}
\usage{
assertListOfPointSets(x)
}
\arguments{
\item{x}{[list]
List of sets, i.e., matrizes.}
}
\value{
Nothing
}
\description{
Helper function to check if all given sets have the same dimension, i.e.,
  number of objectives.
}

