% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assertSameDimensions}
\alias{assertSameDimensions}
\title{Assert same dimensions.}
\usage{
assertSameDimensions(...)
}
\arguments{
\item{...}{[any]
Vectors or matrizes.}
}
\value{
Nothing. Stops if not all dimensions are equal.
}
\description{
We frequently need to check whether the dimension of an approximation set is
  equal to the dimension of the ideal point and/or nadir point. This does exactly
  that assertion for an arbitrary number of arguments.
}

