% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplotECRResult.R
\name{autoplot.ecr_result}
\alias{autoplot.ecr_result}
\title{Plot optimization trace.}
\usage{
\method{autoplot}{ecr_result}(object, xlim = NULL, ylim = NULL,
  log.fitness = FALSE, ...)
}
\arguments{
\item{object}{[\code{ecr_result}]\cr
ecr result object.}

\item{xlim}{[\code{numeric(2)} | NULL]\cr
Lower and upper bound for generation. If \code{NULL}, this is set automatically.}

\item{ylim}{[\code{numeric(2)} | NULL]\cr
Lower and upper bound for fitness. If \code{NULL}, this is set automatically.}

\item{log.fitness}{[\code{logical(1)}]\cr
Log-transform fitness values? Default is \code{FALSE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
}
\description{
Plot optimization trace.
}

