% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutator.bitflip.R
\docType{data}
\name{bitflipMutator}
\alias{bitflipMutator}
\title{Bitplip mutation operator.}
\format{\preformatted{function (setOfIndividuals, control)  
 - attr(*, "name")= chr "Bitplip mutator"
 - attr(*, "description")= chr "Flips each bit of the allele with a specific probability."
 - attr(*, "supported")= chr "binary"
 - attr(*, "class")= chr [1:2] "ecr_operator" "ecr_mutator"
 - attr(*, "defaults")=List of 1
  ..$ mutator.flip.prob: num 0.1
 - attr(*, "checkFunction")=function (setOfIndividuals, control)  
  ..- attr(*, "name")= chr "Bitplip mutator"
  ..- attr(*, "description")= chr "Flips each bit of the allele with a specific probability."
  ..- attr(*, "supported")= chr "binary"
  ..- attr(*, "class")= chr [1:2] "ecr_operator" "ecr_mutator"
  ..- attr(*, "defaults")=List of 1
  .. ..$ mutator.flip.prob: num 0.1
}}
\usage{
bitflipMutator
}
\arguments{
\item{setOfIndividuals}{[\code{setOfIndividuals}]\cr
Set of individuals.}

\item{control}{[\code{list}]\cr
Not used.}
}
\value{
[\code{setOfIndividuals}]
  Set of individuals.
}
\description{
Bitplip mutation operator.
}
\keyword{datasets}

