% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTheEvolution.R
\name{buildInitialPopulation}
\alias{buildInitialPopulation}
\title{Helper function to build initial population.}
\usage{
buildInitialPopulation(n.population, task, control, initial.population = NULL)
}
\arguments{
\item{n.population}{[\code{integer(1)}]\cr
Size of the population.}

\item{task}{[\code{ecr_optimization_task}]\cr
Optimization task.}

\item{control}{[\code{ecr_control}]\cr
Control object.}

\item{initial.population}{[\code{list} | \code{NULL}]\cr
Eventually a list of initial individuals.}
}
\value{
[\code{ecr_population}]
}
\description{
Generates the initial population of an EA based on a list of initial individuals
and/or the control object and the corresponding population generator.
}

