% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeAverageHausdorffDistance.R
\name{computeAverageHausdorffDistance}
\alias{computeAverageHausdorffDistance}
\title{Computes the Average Hausdorff Distance.}
\usage{
computeAverageHausdorffDistance(A, B, p = 1)
}
\arguments{
\item{A}{[\code{matrix}]\cr
First point set (each column corresponds to a point).}

\item{B}{[\code{matrix}]\cr
Second point set (each column corresponds to a point).}

\item{p}{[\code{numeric(1)}]\cr
Parameter p of the average Hausdoff metrix. Default is 1. See the description
for details.}
}
\value{
[\code{numeric(1)}] Average Hausdorff distance of sets \code{A} and \code{B}.
}
\description{
Computes the Average Hausdorff Distance.
}

