% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeCrowdingDistance.R
\name{computeCrowdingDistance}
\alias{computeCrowdingDistance}
\title{Compute the crowding distance of a set of points.}
\usage{
computeCrowdingDistance(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric matrix with each row represneting a points.}
}
\value{
[\code{numeric}] Vector of crowding distance values.
}
\description{
Compute the crowding distance of a set of points.
}
\references{
K. Deb, A. Pratap, S. Agarwal, T. Meyarivan, A fast and elitist
multiobjective genetic algorithm: NSGA-II, IEEE Transactions on Evolutionary
Computation In Evolutionary Computation, IEEE Transactions on, Vol. 6, No. 2.
(07 April 2002), pp. 182-197, doi:10.1109/4235.996017
}

