% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeDominatedHypervolume.R
\name{computeDominatedHypervolume}
\alias{computeDominatedHypervolume}
\title{Computation of the dominated hypervolume.}
\usage{
computeDominatedHypervolume(x, ref.point = NULL)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Matrix of points.}

\item{ref.point}{[\code{numeric} | \code{NULL}]\cr
Reference point. Set to the maximum in each dimension by default if not provided.}
}
\value{
[\code{numeric(1)}] Dominated hypervolume.
}
\description{
Given a set of points \code{x}, this function computes the dominated hypervolume
of the points regarding the reference point \code{ref.point}. If the latter
is not provided, one is automatically determined by computing the maximum
in each dimension.
}

