% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeEpsilonIndicator.R
\name{computeEpsilonIndicator}
\alias{computeEpsilonIndicator}
\title{Computation of the unary epsilon-indicator.}
\usage{
computeEpsilonIndicator(x, ref.points)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Matrix of points.}

\item{ref.points}{[\code{matrix}]\cr
Set of reference points.}
}
\value{
[\code{numeric(1)}] Epsilon indicator.
}
\description{
Given a set of points \code{x}, this function computes the unary epsilon-indicator
provided a set of reference points \code{ref.points}.
}

