% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeFitness.R
\name{computeFitness}
\alias{computeFitness}
\title{Computes the fitness values for each member of a given population.}
\usage{
computeFitness(population, fitness.fun)
}
\arguments{
\item{population}{[\code{matrix}]\cr
Population.}

\item{fitness.fun}{[\code{function}]\cr
Fitness function.}
}
\value{
[\code{matrix}].
}
\description{
This function expects a population, computes the fitness and
  returns the matrix of fitness values.
}

