% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeAverageHausdorffDistance.R
\name{computeInvertedGenerationalDistance}
\alias{computeInvertedGenerationalDistance}
\title{Computes Inverted Generational Distance.}
\usage{
computeInvertedGenerationalDistance(A, B, p = 1)
}
\arguments{
\item{A}{[\code{matrix}]\cr
First point set (each row corresponds to a point).}

\item{B}{[\code{matrix}]\cr
Second point set (each row corresponds to a point).}

\item{p}{[\code{numeric(1)}]\cr
Parameter p of the average Hausdoff metrix. Default is 1. See the description
for details.}
}
\value{
[\code{numeric(1)}]
}
\description{
Helper to compute the Inverted Generational Distance (IGD) between two sets
  of points.
}

