% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recombinator.crossover.R
\docType{data}
\name{crossoverRecombinator}
\alias{crossoverRecombinator}
\title{Generator of the crossover recombination operator.}
\format{\preformatted{function (setOfIndividuals, control = list())  
 - attr(*, "name")= chr "Crossover recombinator"
 - attr(*, "description")= chr "No description"
 - attr(*, "supported")= chr [1:2] "float" "binary"
 - attr(*, "n.parents")= int 2
 - attr(*, "defaults")= list()
 - attr(*, "class")= chr [1:2] "ecr_operator" "ecr_recombinator"
 - attr(*, "checkFunction")=function (operator.control)  
}}
\usage{
crossoverRecombinator
}
\arguments{
\item{setOfIndividuals}{[\code{setOfIndividuals}]\cr
Set of individuals.}

\item{control}{[\code{list}]\cr
Empty list. Intermediate recombinator has no parameters.}
}
\value{
[\code{setOfIndividuals}]
  Recombined offspring.
}
\description{
Generator of the crossover recombination operator.
}
\keyword{datasets}

