% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dominates.R
\name{dominated}
\alias{dominated}
\title{Check for pareto dominance.}
\usage{
dominated(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric (n x d) matrix where n is the number of points and d is the number
of objectives.}
}
\value{
[\code{logical}]
}
\description{
This function takes a numeric matrix \code{x} where each row corresponds to
a point and returns a logical vector. The i-th position of the latter is
\code{TRUE} if the i-th point is dominated by at least one other point.
}

