% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominates.R
\name{dominates}
\alias{dominates}
\title{Check if a vector dominates another vector.}
\usage{
dominates(x, y)
}
\arguments{
\item{x}{[\code{numeric}]\cr
First vector.}

\item{y}{[\code{numeric}]\cr
Second vector.}
}
\value{
[\code{logical(1)}] Does \code{x} dominate \code{y}?
}
\keyword{optimize}

