% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doNondominatedSorting.R
\name{doNondominatedSorting}
\alias{doNondominatedSorting}
\title{Non-dominated sorting algorithm.}
\usage{
doNondominatedSorting(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric matrix of points. Each row contains one objective vector.}
}
\value{
[\code{list}]
  List with the following components
  \describe{
    \item{ranks}{Integer vector of ranks of length \code{nrow(x)}. The higher
    the rank, the higher the domination front the corresponding points is
    located on.}
    \item{dom.counter}{Integer vector of length \code{nrow(x)}. The i-th element
    is the domination counter / domination number of the i-th point.}
  }
}
\description{
Implementation of the fast non-dominated sorting algorithm proposed by Deb.
}

