% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTerminate.R
\name{doTerminate}
\alias{doTerminate}
\title{Check termination conditions.}
\usage{
doTerminate(stopping.funs, opt.path)
}
\arguments{
\item{stopping.funs}{[list]
List of stopping conditions.}

\item{opt.path}{[OptPath]
Optimization path.}
}
\value{
[stopObject]
}
\description{
Helper function which checks whether some stopping conditions
  are met and returns a termination object.
}

