% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupResult.R
\name{ecr_result}
\alias{ecr_result}
\title{Single objective result object.}
\description{
Object returned by \code{\link{doTheEvolution}} in case of the objective function
being single-objective.
}
\details{
The S3 object containts the following members:
\describe{
  \item{task}{The \code{ecr_optimization_task}.}
  \item{control}{The \code{ecr_control} object passed to \code{\link{doTheEvolution}}.}
  \item{best.param}{Overall best parameters.}
  \item{best.value}{Overall best objective value.}
  \item{opt.path}{Optimization path \code{\link[ParamHelpers]{OptPath}}.}
  \item{population.storage}{Named list of populations stored during the process.}
  \item{message}{Character string describing the reason of termination.}
}
}

