% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeECRResult.R
\name{ECRMultiObjectiveResult}
\alias{ECRMultiObjectiveResult}
\title{Multi objective result object.}
\description{
Object returned by \code{\link{doTheEvolution}} in case of the objective function
being single-objective.
}
\details{
It contains ...
}

