% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeECRResult.R
\name{ECRSingleObjectiveResult}
\alias{ECRSingleObjectiveResult}
\title{Single objective result object.}
\description{
Object returned by \code{\link{doTheEvolution}} in case of the objective function
being single-objective.
}
\details{
It contains ...
}

