% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupEvolutionaryOperators.R
\name{setupParentSelector}
\alias{setupGenerator}
\alias{setupMutator}
\alias{setupParentSelector}
\alias{setupRecombinator}
\alias{setupSurvivalSelector}
\title{Collection of functions to set specific evolutionay operators.}
\usage{
setupParentSelector(control, operator)

setupSurvivalSelector(control, operator)

setupGenerator(control, operator)

setupMutator(control, operator)

setupRecombinator(control, operator)
}
\arguments{
\item{control}{[\code{ecr_control}]\cr
ECR control object generated via \code{\link{setupECRControl}}.}

\item{operator}{[\code{ecr_operator}]\cr
The matching ecr operator, e.g., an object of type \code{ecr_mutator} for
\code{\link{setupMutator}}.}
}
\description{
This functions can be used to append/set specific evolutionary operators
to the control object. However, multiple operators can be set with one function
call via \code{\link{setupEvolutionaryOperators}}.
}

