% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateOffspring.R
\name{generateOffspring}
\alias{generateOffspring}
\title{Creates offspring from a given mating pool of parents.}
\usage{
generateOffspring(matingPool, STORAGE, objective.fun, control, opt.path)
}
\arguments{
\item{matingPool}{[\code{setOfIndividuals}]\cr
Set of parents to choose from.}

\item{STORAGE}{[\code{list}]\cr
List which contains all the algorithm specific stuff.}

\item{objective.fun}{[\code{function}]\cr
Target fun.}

\item{control}{[\code{ecr_control}]\cr
Control object containing alle the operators and further parameters.}

\item{opt.path}{[\code{OptPath}]\cr
Optimization path.}
}
\value{
[\code{setOfIndividuals}] Generated offspring.
}
\description{
Creates offspring from a given mating pool of parents.
}

