% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDispatcher.R
\name{getAvailableEventNames}
\alias{getAvailableEventNames}
\title{Receive a character vector of all valid event names.}
\usage{
getAvailableEventNames()
}
\description{
Receive a character vector of all valid event names.
}

