% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{getDefaults}
\alias{getDefaults}
\title{Get a list of the default operator parameters.}
\usage{
getDefaults(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator.}
}
\value{
[\code{list}]
}
\description{
Get a list of the default operator parameters.
}

