% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeECRResult.R
\name{getGenerations}
\alias{getGenerations}
\title{Determine number of generations.}
\usage{
getGenerations(result)
}
\arguments{
\item{result}{[\code{ecr_result}]\cr
\pkg{ecr} result object.}
}
\value{
[integer(1)]
}
\description{
Determine number of generations.
}

