% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNextGeneration.R
\name{getNextGeneration}
\alias{getNextGeneration}
\title{Performs survival selection.}
\usage{
getNextGeneration(population, offspring, STORAGE, n.population,
  strategy = "plus", n.elite = 0L, control)
}
\arguments{
\item{population}{[\code{setOfIndividuals}]\cr
Population.}

\item{offspring}{[\code{integer(1)}]\cr
Generated offspring.}

\item{STORAGE}{[\code{list}]\cr
List which contains all the algorithm specific stuff.}

\item{n.population}{[\code{integer(1)}]\cr
Number of individuals.}

\item{strategy}{[\code{character(1)}]\cr
Strategy used for selection. Possible strategies are:
\describe{
  \item{plus}{A classical (mu + lambda) strategy.}
  \item{comma}{A classical (mu, lambda) strategy.}
}
Default is \code{plus}. Another is not implemented yet.}

\item{n.elite}{[\code{integer(1)}]\cr
Number of fittest individuals of the current generation that shall be copied to the
next generation without changing. Default is 0.}

\item{control}{[\code{ecr_control}]\cr
Control object.}
}
\value{
[\code{setOfIndividuals}]
}
\description{
Helper function which selects the individuals of a population which
  will survive the current generation.
}

