% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{getOperatorCheckFunction}
\alias{getOperatorCheckFunction}
\title{Returns the parameter check function of a given operator.}
\usage{
getOperatorCheckFunction(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{character(1)}]
  Function for parameter check.
}

