% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{getOperatorParameters}
\alias{getOperatorParameters}
\title{Return a list of parameters the operator was initialized with.}
\usage{
getOperatorParameters(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{list}]
  Named list of parameters.
}

