% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateOffspring.R
\name{getParents}
\alias{getParents}
\title{Helper method to extract two parents from the mating pool}
\usage{
getParents(matingPool)
}
\arguments{
\item{matingPool}{[ecr_population]
Set of individuals selected for reproduction.}
}
\value{
[list]
}
\description{
Helper method to extract two parents from the mating pool
}

