% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSupportedRepresentations.R
\name{getSupportedRepresentations}
\alias{getSupportedRepresentations}
\title{Returns the representations which a specific operator supports.}
\usage{
getSupportedRepresentations(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{character}]
  Vector of representation types.
}
\description{
Operators and representation types are not mandatory compatible. This function
detemines the types an operator can operate on.
}

