% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{hasAllEqualElements}
\alias{hasAllEqualElements}
\title{Check whether all elements of an vector are equal.}
\usage{
hasAllEqualElements(x)
}
\arguments{
\item{x}{[vector]
Input vector.}
}
\value{
[logical(1)]
}
\description{
Check whether all elements of an vector are equal.
}

