\name{intermediateRecombinator}
\alias{intermediateRecombinator}
\title{Generator of the indermediate recombination operator.}
\usage{
intermediateRecombinator(setOfIndividuals, control = list())
}
\arguments{
  \item{individuals}{[\code{setOfIndividuals}]\cr Set of
  individuals.}

  \item{control}{[\code{list}]\cr Empty list. Intermediate
  recombinator has no paramerters.}
}
\value{
[\code{setOfIndividuals}] Recombined offspring.
}
\description{
Generator of the indermediate recombination operator.
}

