\name{is.supported}
\alias{is.supported}
\title{Helper function to check if esoo operator supports given representation.}
\usage{
is.supported(operator, representation)
}
\arguments{
  \item{operator}{[\code{esoo_operator}]\cr Object of type
  \code{esoo_operator}.}

  \item{representation}{[\code{character(1)}]\cr
  Representation as a string.}
}
\value{
[\code{logical(1)}] \code{TRUE}, if operator supports the
representation type.
}
\description{
Helper function to check if esoo operator supports given
representation.
}

