% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dominates.R
\name{isDominated}
\alias{isDominated}
\title{Check whether a vector is dominated by another vector.}
\usage{
isDominated(x, y)
}
\arguments{
\item{x}{[\code{numeric}]\cr
First vector.}

\item{y}{[\code{numeric}]\cr
Second vector.}
}
\value{
[\code{logical(1)}] Returns \code{TRUE} if \code{x} is dominated by
  \code{y} and \code{FALSE} otherwise.
}
\description{
Check whether a vector is dominated by another vector.
}
\keyword{optimize}

