% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generator.random.binary.R
\name{makeBinaryGenerator}
\alias{makeBinaryGenerator}
\title{Generates a generator object for the initial population.}
\usage{
makeBinaryGenerator()
}
\value{
[\code{ecr_operator}]
}
\description{
Generates a generator object for the initial population.
}

