% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutator.bitflip.R
\name{makeBitFlipMutator}
\alias{makeBitFlipMutator}
\title{Bitplip mutation operator.}
\usage{
makeBitFlipMutator(mutator.flip.prob = 0.1)
}
\arguments{
\item{mutator.flip.prob}{[\code{numeric(1)}]\cr
Probability to flip a single bit. Default is \code{0.1}.}
}
\value{
[\code{ecr_mutator}]
}
\description{
Bitplip mutation operator.
}

