% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingCondition.optimum.R
\name{makeCloseToOptimumStoppingCondition}
\alias{makeCloseToOptimumStoppingCondition}
\title{Stopping condition: close to optimum.}
\usage{
makeCloseToOptimumStoppingCondition(eps = 1e-04, opt = NULL)
}
\arguments{
\item{eps}{[\code{numeric(1)}]\cr
Maximal deviation from optimal value.}

\item{opt}{[\code{numeric(1)}]\cr
Optimal value.}
}
\value{
[\code{function}]
}
\description{
Stop the EA if the best fitness value is close to the given optimum.
}
\seealso{
Other stopping conditions: \code{\link{makeMaximumEvaluationsStoppingCondition}},
  \code{\link{makeMaximumIterationsStoppingCondition}},
  \code{\link{makeMaximumTimeStoppingCondition}}
}

