% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombinator.crossover.R
\name{makeCrossoverRecombinator}
\alias{makeCrossoverRecombinator}
\title{Generator of the One-point crossover recombination operator.}
\usage{
makeCrossoverRecombinator()
}
\value{
[\code{ecr_recombinator}]
}
\description{
Generator of the One-point crossover recombination operator.
}
\seealso{
Other recombinators: \code{\link{makeIntermediateRecombinator}};
  \code{\link{makeNullRecombinator}};
  \code{\link{makeOXRecombinator}};
  \code{\link{makePMXRecombinator}}
}

