% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutator.gauss.R
\name{makeGaussMutator}
\alias{makeGaussMutator}
\title{Generator of the Gauss mutation operator.}
\usage{
makeGaussMutator(p = 1L, sdev = 0.05)
}
\arguments{
\item{p}{[\code{numeric(1)}]\cr
Probability of mutation for the gauss mutation operator.}

\item{sdev}{[\code{numeric(1)}\cr
Standard deviance of the Gauss mutation, i. e., the mutation strength.}
}
\value{
[\code{ecr_mutator}]
}
\description{
Default Gaussian mutation operator known from Evolutionary Algorithms.
}

