% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutator.insertion.R
\name{makeInsertionMutator}
\alias{makeInsertionMutator}
\title{Generator for the Insertion mutation operator.}
\usage{
makeInsertionMutator()
}
\value{
[\code{ecr_mutator}]
}
\description{
The Insertion mutation operator selects a position random and inserts it at
a random position.
}

