% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinator.intermediate.R
\name{makeIntermediateRecombinator}
\alias{makeIntermediateRecombinator}
\title{Generator of the indermediate recombination operator.}
\usage{
makeIntermediateRecombinator(k = 2L)
}
\arguments{
\item{k}{[integer(1)]\cr
Number of parents required for mating. Default is \code{2}.}
}
\value{
[\code{ecr_recombinator}]
}
\seealso{
Other recombinators: \code{\link{makeCrossoverRecombinator}},
  \code{\link{makeNullRecombinator}},
  \code{\link{makeOXRecombinator}},
  \code{\link{makePMXRecombinator}}
}

