% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.inversion.R
\name{makeInversionMutator}
\alias{makeInversionMutator}
\title{Generator for the Inversion mutation operator.}
\usage{
makeInversionMutator()
}
\value{
[\code{ecr_mutator}]
}
\description{
The Inversion mutation operator selects two positions within the chromosome at
random and inverts this sub-permutation.
}

