% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingCondition.max.evaluations.R
\name{makeMaximumEvaluationsStoppingCondition}
\alias{makeMaximumEvaluationsStoppingCondition}
\title{Stopping condition: maximum number of function evaluations.}
\usage{
makeMaximumEvaluationsStoppingCondition(max.evals = NULL)
}
\arguments{
\item{max.evals}{[\code{integer(1)}]\cr
Maximal number of function evaluations. Default ist \code{Inf}.}
}
\value{
[\code{function}]
}
\description{
Stopping condition: maximum number of function evaluations.
}

