% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingCondition.max.iter.R
\name{makeMaximumIterationsStoppingCondition}
\alias{makeMaximumIterationsStoppingCondition}
\title{Stopping condition: maximum number of iterations.}
\usage{
makeMaximumIterationsStoppingCondition(max.iter = NULL)
}
\arguments{
\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations. Default ist \code{Inf}.}
}
\value{
[\code{function}]
}
\description{
Stopping condition: maximum number of iterations.
}

