% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingCondition.max.time.R
\name{makeMaximumTimeStoppingCondition}
\alias{makeMaximumTimeStoppingCondition}
\title{Stopping condition: time limit.}
\usage{
makeMaximumTimeStoppingCondition(max.time = NULL)
}
\arguments{
\item{max.time}{[\code{integer(1)}]\cr
Time budget in seconds. Default ist \code{Inf}.}
}
\value{
[\code{function}]
}
\description{
Stop the EA after a maximal time budget.
}
\seealso{
Other stopping conditions: \code{\link{makeCloseToOptimumStoppingCondition}},
  \code{\link{makeMaximumEvaluationsStoppingCondition}},
  \code{\link{makeMaximumIterationsStoppingCondition}}
}

