% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombinator.null.R
\name{makeNullRecombinator}
\alias{makeNullRecombinator}
\title{This recombinator does not perform any recombination.}
\usage{
makeNullRecombinator()
}
\value{
[\code{ecr_recombinator}]
}
\description{
This recombinator does not perform any recombination.
}
\seealso{
Other recombinators: \code{\link{makeCrossoverRecombinator}};
  \code{\link{makeIntermediateRecombinator}};
  \code{\link{makeOXRecombinator}};
  \code{\link{makePMXRecombinator}}
}

