% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{makeOperator}
\alias{makeOperator}
\title{Construct evolutionary operator.}
\usage{
makeOperator(operator, name, description,
  supported = getAvailableRepresentations(), defaults = list(),
  checker = function(operator.control) TRUE)
}
\arguments{
\item{operator}{[\code{function}]\cr
Actual mutation operator.}

\item{name}{[\code{character(1)}]\cr
Name of the operator.}

\item{description}{[\code{character(1)}]\cr
Short description of how the mutator works.}

\item{supported}{[\code{character}]\cr
Vector of names of supported parameter representations. For example
'permutation', 'float', 'binary'.}

\item{defaults}{[\code{list}]\cr
List of default values for the operators strategy parameters.}

\item{checker}{[\code{function}]\cr
Check object, which performs a sanity check of the strategy parameters
passed to the control object.
Operator object.}
}
\value{
[\code{ecr_operator}]
}
\description{
Helper function which constructs an evolutionary operator.
}
\note{
In general you will not need this function, but rather one of its
  deriviatives like \code{\link{makeMutator}} or \code{\link{makeSelector}}.
}

