% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.random.permutation.R
\name{makePermutationGenerator}
\alias{makePermutationGenerator}
\title{Generates a generator object for the initial population.}
\usage{
makePermutationGenerator()
}
\value{
[\code{ecr_operator}]
}
\description{
This generator generates a population of permutations.
}

