% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.polynomial.R
\name{makePolynomialMutation}
\alias{makePolynomialMutation}
\title{Generator of the polynomial mutation operator.}
\usage{
makePolynomialMutation(p = 0.2, eta = 10)
}
\arguments{
\item{p}{[\code{numeric(1)}]\cr
Probability of mutation of each gene.}

\item{eta}{[\code{numeric(1)}\cr
Distance parameter of the mutation distribution.}
}
\value{
[\code{ecr_mutator}]
}
\description{
Performs an polynomial mutation as used in the SMS-EMOA algorithm.
}
\seealso{
Other mutators: \code{\link{makeBitFlipMutator}},
  \code{\link{makeGaussMutator}},
  \code{\link{makeInsertionMutator}},
  \code{\link{makeScrambleMutator}},
  \code{\link{makeSwapMutator}},
  \code{\link{makeUniformMutator}}
}

