% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.roulettewheel.R
\name{makeRouletteWheelSelector}
\alias{makeRouletteWheelSelector}
\title{Generator for Roulette-Wheel-Selection (fitness-proportional selection).}
\usage{
makeRouletteWheelSelector()
}
\value{
[\code{setOfIndividuals}]
}
\description{
Selects genes for the mating pool. The chance of an individual to get selected
  is proportional to its fitness, i.e., better individuals get a higher chance
  to take part in the reproduction process. Low-fitness individuals however,
  have a positive fitness as well.
}
\seealso{
Other selectors: \code{\link{makeGreedySelector}},
  \code{\link{makeSimpleSelector}},
  \code{\link{makeTournamentSelector}}
}

