% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinator.sbx.R
\name{makeSBXRecombinator}
\alias{makeSBXRecombinator}
\title{Generator of the Simulated Binary Crossover (SBX) recombinator.}
\usage{
makeSBXRecombinator(eta = 5, p = 1)
}
\arguments{
\item{eta}{[\code{numeric(1)}]\cr
Parameter eta, i.e., the distance parameters of the crossover distribution.}

\item{p}{[\code{numeric(1)}]\cr
Crossover probability for each gene. Default is \code{1.0}.}
}
\value{
[\code{ecr_recombinator}]
}
\description{
...
}
\seealso{
Other recombinators: \code{\link{makeCrossoverRecombinator}},
  \code{\link{makeIntermediateRecombinator}},
  \code{\link{makeNullRecombinator}},
  \code{\link{makeOXRecombinator}},
  \code{\link{makePMXRecombinator}}
}

