% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.simple.R
\name{makeSimpleSelector}
\alias{makeSimpleSelector}
\title{Simple (naive) mating pool generator.}
\usage{
makeSimpleSelector()
}
\value{
[\code{setOfIndividuals}]
}
\description{
Just for testing. Actually does not really select, but instead returns the
  entire population to form the mating pool.
}
\seealso{
Other selectors: \code{\link{makeGreedySelector}},
  \code{\link{makeRouletteWheelSelector}},
  \code{\link{makeTournamentSelector}}
}

