% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutator.swap.R
\name{makeSwapMutator}
\alias{makeSwapMutator}
\title{Generator for the Swap mutation operator.}
\usage{
makeSwapMutator()
}
\value{
[\code{ecr_mutator}]
}
\description{
Chooses two positions at random and swaps the genes.
}
\seealso{
Other mutators: \code{\link{makeBitFlipMutator}};
  \code{\link{makeGaussMutator}};
  \code{\link{makeInsertionMutator}};
  \code{\link{makeScrambleMutator}};
  \code{\link{makeUniformMutator}}
}

