% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.uniform.R
\name{makeUniformMutator}
\alias{makeUniformMutator}
\title{Generator of the Uniform mutation operator.}
\usage{
makeUniformMutator()
}
\value{
[\code{ecr_mutator}]
}
\description{
This mutation operator works on real-valued genotypes only. It selects a position
in the solution vector at random and replaced it with a uniformally chosen value
within the box constraints of the corresponding parameter.
This mutator may proof beneficial in early stages of the optimization process,
since it distributes points widely within the box constraints and thus may
hinder premature convergence. However, in later stages - when fine tuning is
necessary, this feature is disadvantegous.
}
\seealso{
Other mutators: \code{\link{makeBitFlipMutator}},
  \code{\link{makeGaussMutator}},
  \code{\link{makeInsertionMutator}},
  \code{\link{makePolynomialMutation}},
  \code{\link{makeScrambleMutator}},
  \code{\link{makeSwapMutator}}
}

