% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptState.getter.R
\name{OptStateGetter}
\alias{OptStateGetter}
\alias{getOptStateBestIndividual}
\alias{getOptStateControl}
\alias{getOptStateCurrentEvaluations}
\alias{getOptStateCurrentIter}
\alias{getOptStateFitness}
\alias{getOptStateOptPath}
\alias{getOptStateParamSet}
\alias{getOptStatePopulation}
\alias{getOptStateTask}
\alias{getOptStateTimePassed}
\title{Optimization state getter functions.}
\usage{
getOptStateCurrentIter(opt.state)

getOptStateTask(opt.state)

getOptStateParamSet(opt.state)

getOptStateCurrentEvaluations(opt.state)

getOptStateControl(opt.state)

getOptStateBestIndividual(opt.state)

getOptStatePopulation(opt.state)

getOptStateFitness(opt.state)

getOptStateTimePassed(opt.state)

getOptStateOptPath(opt.state)
}
\arguments{
\item{opt.state}{[\code{\link{OptState}}]\cr
Optimization state.}
}
\value{
[any] Type depends on the property.
}
\description{
The \code{\link{OptState}} is generated internally by \code{\link{doTheEvolution}},
but it is exposed to the user in some cases. E.g., self-defined actions or
stopping conditions get the current \code{\link{OptState}} passed as an argument.
The following collection of simple getter functions helps to access the
properties of the optimization state.
}

