% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.ecr_control}
\alias{print.ecr_control}
\title{Print ecr control object.}
\usage{
\method{print}{ecr_control}(x, ...)
}
\arguments{
\item{x}{[\code{ecr_control}]\cr
Control object.}

\item{...}{[any]\cr
  Not used.}
}
\description{
Print ecr control object.
}

