% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doTheEvolution.R
\name{print.ecr_result}
\alias{print.ecr_result}
\title{Print the result of an ecr run.}
\usage{
\method{print}{ecr_result}(x, ...)
}
\arguments{
\item{x}{[\code{ecr_result}]\cr
ecr result object.}

\item{...}{[any]\cr
Not used.}
}
\description{
Print the result of an ecr run.
}

