% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDispatcher.R
\name{registerAction}
\alias{registerAction}
\title{Register an action.}
\usage{
registerAction(control, event.name, fun)
}
\arguments{
\item{control}{[\code{ecr_control}]\cr
Control object.}

\item{event.name}{[\code{character(1)}]\cr
Name of the event.}

\item{fun}{[\code{function()}]\cr
Function with arguments \code{opt.state} and \code{...}.}
}
\value{
[\code{ecr_control}] Updated control object.
}
\description{
Register an action for a specific event in the EA circle.
See \code{\link{getAvailableEventNames}} for a list of available event names.
}
\seealso{
\code{\link{getAvailableEventNames}}
}

