% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupECRControl.R
\name{selectForMating}
\alias{selectForMating}
\title{Generate mating pool.}
\usage{
selectForMating(opt.state, control)
}
\arguments{
\item{opt.state}{[\code{ecr_opt_state}]\cr
Optimization state.}

\item{control}{[\code{ecr_control}]\cr
ECR control object generated via \code{\link{setupECRControl}}.}
}
\value{
[\code{ecr_population}]
}
\description{
Given the current optimization state this function selects a subset of individuals
which form the mating pool, i.e., the set of possible parents for later recombination.
}
\note{
Basically this is a wrapper for the call to the parent selector of the control
object which occasionally transforms the fitness values (e.g., if maximzation should be
performed, but the selection operator actualy minimizes).
}

