% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.terminator.optimum.R
\name{setupCloseToOptimumTerminator}
\alias{setupCloseToOptimumTerminator}
\title{Stopping condition: close to optimum.}
\usage{
setupCloseToOptimumTerminator(eps = 1e-04, opt = NULL)
}
\arguments{
\item{eps}{[\code{numeric(1)}]\cr
Maximal deviation from optimal value.}

\item{opt}{[\code{numeric(1)}]\cr
Optimal value.}
}
\value{
[\code{function}]
}
\description{
Stop the EA if the best fitness value is close to the given optimum.
}
\seealso{
Other stopping conditions: \code{\link{setupMaximumEvaluationsTerminator}},
  \code{\link{setupMaximumIterationsTerminator}},
  \code{\link{setupMaximumTimeTerminator}}
}

