% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupECRControl.R
\name{setupECRControl}
\alias{setupECRControl}
\title{Generates control object.}
\usage{
setupECRControl(n.population, n.offspring,
  n.mating.pool = floor(n.population/2), representation,
  survival.strategy = "plus", n.elite = 0L, target.name = "y",
  save.population.at = integer(0), monitor = makeConsoleMonitor(),
  stopping.conditions = list(), extras.fun = NULL)
}
\arguments{
\item{n.population}{[\code{integer(1)}]\cr
Number of individuals in the population.}

\item{n.offspring}{[\code{integer(1)}]\cr
Number of individuals generated in each generation.}

\item{n.mating.pool}{[\code{integer(1)}]\cr
Number of individuals which can potentially participate in the
generation of offspring. Default is half of the population size.}

\item{representation}{[\code{character(1)}]\cr
Genotype representation of the parameters. Available are binary, real,
permutation and custom.}

\item{survival.strategy}{[\code{character(1)}]\cr
Determines the survival strategy used by the EA. Possible are 'plus' for a classical
(mu + lambda) strategy and 'comma' for (mu, lambda).}

\item{n.elite}{[\code{integer(1)}]\cr
Number of fittest individuals of the current generation that shall be copied to the
next generation without changing. Default is 0. Keep in mind, that the algorithm
does not care about this option if the \code{survival.strategy} is set to 'plus'.}

\item{target.name}{[\code{character(1)}]\cr
Name for the objective fun values. Default is \dQuote{y}.}

\item{save.population.at}{[\code{integer}]\cr
Which populations should be saved? Default is none.}

\item{monitor}{[\code{function}]\cr
Monitoring function. Default is \code{consoleMonitor}.}

\item{stopping.conditions}{[\code{list}]\cr
List of functions of type \code{ecr_stoppingCondition}.}

\item{extras.fun}{[\code{function} | \code{NULL}]\cr
Functin which expects the population and returns a list of scalar quality
computed based on the individual genomes or the fitness values, e.g., standard
deviation of the fitness. The function is called once in each generation.
The results are stored in the optimization path. Default is \code{NULL}, which
means that no additional stuff is logged.}
}
\value{
S3 object of type \code{ecr_control}.
}
\description{
The ecr package offers a framework for evolutionary computing and therefore offers
  a lot of customization options. The control object is a simple but powerful
  wrapper for all these options and sets convenient default options.
}

