% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.gauss.R
\name{setupGaussMutator}
\alias{setupGaussMutator}
\title{Generator of the Gauss mutation operator.}
\usage{
setupGaussMutator(p = 1L, sdev = 0.05)
}
\arguments{
\item{p}{[\code{numeric(1)}]\cr
Probability of mutation for the gauss mutation operator.}

\item{sdev}{[\code{numeric(1)}\cr
Standard deviance of the Gauss mutation, i. e., the mutation strength.}
}
\value{
[\code{ecr_mutator}]
}
\description{
Default Gaussian mutation operator known from Evolutionary Algorithms.
}
\seealso{
Other mutators: \code{\link{setupBitFlipMutator}},
  \code{\link{setupInsertionMutator}},
  \code{\link{setupPolynomialMutator}},
  \code{\link{setupScrambleMutator}},
  \code{\link{setupSwapMutator}},
  \code{\link{setupUniformMutator}}
}

