% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.terminator.max.evaluations.R
\name{setupMaximumEvaluationsTerminator}
\alias{setupMaximumEvaluationsTerminator}
\title{Stopping condition: maximum number of function evaluations.}
\usage{
setupMaximumEvaluationsTerminator(max.evals = NULL)
}
\arguments{
\item{max.evals}{[\code{integer(1)}]\cr
Maximal number of function evaluations. Default ist \code{Inf}.}
}
\value{
[\code{function}]
}
\description{
Stop the EA after at most \code{max.evals} evaluations of the fitness function.
}
\seealso{
Other stopping conditions: \code{\link{setupCloseToOptimumTerminator}},
  \code{\link{setupMaximumIterationsTerminator}},
  \code{\link{setupMaximumTimeTerminator}}
}

