% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.terminator.max.iter.R
\name{setupMaximumIterationsTerminator}
\alias{setupMaximumIterationsTerminator}
\title{Stopping condition: maximum number of iterations.}
\usage{
setupMaximumIterationsTerminator(max.iter = NULL)
}
\arguments{
\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations. Default ist \code{Inf}.}
}
\value{
[\code{function}]
}
\description{
Stop the EA after at most \code{max.iter} generations/iterations.
}
\seealso{
Other stopping conditions: \code{\link{setupCloseToOptimumTerminator}},
  \code{\link{setupMaximumEvaluationsTerminator}},
  \code{\link{setupMaximumTimeTerminator}}
}

