% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.termnator.max.time.R
\name{setupMaximumTimeTerminator}
\alias{setupMaximumTimeTerminator}
\title{Stopping condition: time limit.}
\usage{
setupMaximumTimeTerminator(max.time = NULL)
}
\arguments{
\item{max.time}{[\code{integer(1)}]\cr
Time budget in seconds. Default ist \code{Inf}.}
}
\value{
[\code{function}]
}
\description{
Stop the EA after a maximal time budget.
}
\seealso{
Other stopping conditions: \code{\link{setupCloseToOptimumTerminator}},
  \code{\link{setupMaximumEvaluationsTerminator}},
  \code{\link{setupMaximumIterationsTerminator}}
}

