% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptState.setter.R
\name{setupOptState}
\alias{setupOptState}
\title{Generate an optimization state object.}
\usage{
setupOptState(task, population, control)
}
\arguments{
\item{task}{[\code{ecr_optimization_task}]\cr
Optimization task.}

\item{population}{[\code{ecr_population}]\cr
(Initial) population.}

\item{control}{[\code{ecr_control}]\cr
Control object.}
}
\value{
[\code{ecr_opt_state}]
}
\description{
The optimization task object is a container for all variables
regarding the optimization process.
}

