% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.generator.random.permutation.R
\name{setupPermutationGenerator}
\alias{setupPermutationGenerator}
\title{Generates a generator object for the initial population.}
\usage{
setupPermutationGenerator()
}
\value{
[\code{ecr_operator}]
}
\description{
This generator generates a population of permutations.
}

