% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.recombinator.sbx.R
\name{setupSBXRecombinator}
\alias{setupSBXRecombinator}
\title{Generator of the Simulated Binary Crossover (SBX) recombinator.}
\usage{
setupSBXRecombinator(eta = 5, p = 1)
}
\arguments{
\item{eta}{[\code{numeric(1)}]\cr
Parameter eta, i.e., the distance parameters of the crossover distribution.}

\item{p}{[\code{numeric(1)}]\cr
Crossover probability for each gene. Default is \code{1.0}.}
}
\value{
[\code{ecr_recombinator}]
}
\description{
...
}
\seealso{
Other recombinators: \code{\link{setupCrossoverRecombinator}},
  \code{\link{setupIntermediateRecombinator}},
  \code{\link{setupNullRecombinator}},
  \code{\link{setupOXRecombinator}},
  \code{\link{setupPMXRecombinator}}
}

