% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupTerminators.R
\name{setupTerminators}
\alias{setupTerminators}
\title{Generator for some frequently used stopping conditions.}
\usage{
setupTerminators(max.iter = NULL, max.time = NULL)
}
\arguments{
\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations. Default ist \code{Inf}.}

\item{max.time}{[\code{integer(1)}]\cr
Time budget in seconds. Default ist \code{Inf}.}
}
\value{
[list]
  List of \code{ecr_terminator} objects, which can be passed to \code{ecr}.
}
\description{
Setting up stopping conditions is flexible and straight-forward, but
it needs a lot of writing. This function generates a list of frequently
used stopping conditions with just a single function call.
}

