% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.generator.uniform.float.R
\name{setupUniformGenerator}
\alias{setupUniformGenerator}
\title{Generates the uniform generator object for the initial population.}
\usage{
setupUniformGenerator()
}
\value{
[\code{ecr_generator}]
}
\description{
The returned population contains individuals which are uniformly distributed
within the bounds specified by the paramter set of the \pkg{smoof} objective
function passed to the \code{\link{doTheEvolution}} function.
}

