% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generator.matingPool.simple.R
\name{simpleMatingPoolGenerator}
\alias{simpleMatingPoolGenerator}
\title{Simple mating pool generator.}
\usage{
simpleMatingPoolGenerator(population, n.mating.pool)
}
\arguments{
\item{population}{[\code{setOfIndividuals}]\cr
Population.}

\item{n.mating.pool}{[\code{integer(1)}]\cr
Number of individuals which should be placed in the mating pool.}
}
\value{
[\code{setOfIndividuals}]
}
\description{
Simple mating pool generator.
}

