% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emoa.sms-emoa.R
\name{smsemoa}
\alias{smsemoa}
\title{Implementation of the SMS-EMOA by Emmerich et al.}
\usage{
smsemoa(task, n.population = 100L, n.offspring = n.population,
  ref.point = NULL, parent.selector = makeSimpleSelector(),
  mutator = makeGaussMutator(), recombinator = makeCrossoverRecombinator(),
  max.iter = 100L, max.evals = NULL, max.time = NULL)
}
\arguments{
\item{task}{[\code{ecr_optimization_task}]\cr
Optimization task. If a \code{smoof_function} is passed it is automatically
converted into a task.}

\item{n.population}{[\code{integer(1)}]\cr
Population size. Default is \code{100}.}

\item{n.offspring}{[\code{integer(1)}]\cr
Offspring size, i.e., number of individuals generated by variation operators
in each iteration. Default is \code{n.population}.}

\item{ref.point}{[\code{numeric}]\cr
Reference point for the hypervolume computation. Default is (11, ..., 11)'
with the corresponding dimension.}

\item{parent.selector}{[\code{ecr_selector}]\cr
Selection operator which implements a procedure to copy individuals from a
given population to the mating pool, i. e., allow them to become parents.}

\item{mutator}{[\code{ecr_mutator}]\cr
Mutation operator of type \code{ecr_mutator}.}

\item{recombinator}{[\code{ecr_recombinator}]\cr
Recombination operator of type \code{ecr_recombinator}.}

\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations. Default ist \code{100L}.}

\item{max.evals}{[\code{integer(1)}]\cr
Maximal number of iterations/generations. Default is \code{Inf}.}

\item{max.time}{[\code{integer(1)}]\cr
Time budget in seconds. Default ist \code{Inf}.}
}
\value{
[\code{ecr_ecr_multi_objective_result}]
}
\description{
Implementation of the SMS-EMOA by Emmerich et al.
}
\note{
This helper function hides the regular \pkg{ecr} interface and offers a more
  R like interface of this state of the art EMOA while still being quite adaptable.
}

