% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{summary.ecr_multi_objective_result}
\alias{summary.ecr_multi_objective_result}
\title{Summary function for multi objective ecr result.}
\usage{
\method{summary}{ecr_multi_objective_result}(object, ref.points = NULL, ...)
}
\arguments{
\item{object}{[\code{ecr_multi_objective_result}]\cr
Result object.}

\item{ref.points}{[\code{matrix}]\cr
Matrix of reference points (one point per column) used for the emoa quality
indicators, i.e., epsilon indicator and hypervolume indicator.}

\item{...}{[any]\cr
Furhter parameters passed to R\{1,2,3\} computation functions. See e.g.
\code{\link{computeR1Indicator}} for details.}
}
\value{
[\code{ecr_multi_objective_result_summary}]
}
\description{
Computes a table of EMOA indicators based on a multi-objective ecr result object.
}

