\name{swapMutator}
\alias{swapMutator}
\title{Swap mutation operator.}
\usage{
swapMutator(setOfIndividuals, control = list())
}
\arguments{
  \item{setOfIndividuals}{[\code{setOfIndividuals}]\cr Set
  of individuals.}

  \item{control}{[\code{list}]\cr Not used.}
}
\value{
[\code{setOfIndividuals}] Set of individuals.
}
\description{
Swap mutation operator.
}

