% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptState.setter.R
\name{updateOptState}
\alias{updateOptState}
\title{Update optimization state.}
\usage{
updateOptState(opt.state, population, control)
}
\arguments{
\item{opt.state}{[\code{ecr_opt_state}]\cr
Optimization state.}

\item{population}{[\code{ecr_population}]\cr
Current population.}

\item{control}{[\code{ecr_control}]\cr.
Control object.}
}
\value{
Nothing.
}
\description{
Called once after survival selection to update the optimization state.
}

