% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrapChildren}
\alias{wrapChildren}
\title{Wrap the individuals constructed by a recombination operator.}
\usage{
wrapChildren(...)
}
\description{
Should be used if the recombinator returns multiple children.
}

