% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAverageHausdorffDistance.R
\name{computeDistanceFromPointToSetOfPoints}
\alias{computeDistanceFromPointToSetOfPoints}
\title{Computes distance between point and set of points.}
\usage{
computeDistanceFromPointToSetOfPoints(a, B)
}
\arguments{
\item{a}{[\code{numeric(1)}]\cr
Point given as a numeric vector.}

\item{B}{[\code{matrix}]\cr
Point set (each row corresponds to a point).}
}
\value{
[\code{numeric(1)}]
}
\description{
Helper to compute distance between a single point and a point set.
}

