% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeFitness.R
\name{computeFitness}
\alias{computeFitness}
\title{Computes the fitness values for each member of a given population.}
\usage{
computeFitness(population, fitness.fun)
}
\arguments{
\item{population}{[\code{matrix}]\cr
Population.}

\item{fitness.fun}{[\code{function}]\cr
Fitness function.}
}
\value{
[\code{matrix}].
}
\description{
This function gets a population computes the fitness and
returns the vector of fitness values.
}

