% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeFitness.R
\name{computeFitness}
\alias{computeFitness}
\title{Computes the fitness values for each member of a given population.}
\usage{
computeFitness(population, fitness.fun, control)
}
\arguments{
\item{population}{[\code{matrix}]\cr
Population.}

\item{fitness.fun}{[\code{function}]\cr
Fitness function.}

\item{control}{[\code{ecr_control}]\cr
Control object containing all operators and further parameters.
See \code{\link{setupECRControl}} and \code{\link{setupEvolutionaryOperators}}.}
}
\value{
[\code{matrix}].
}
\description{
This function gets a population computes the fitness and
returns the vector of fitness values.
}

