% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doNondominatedSorting.R
\name{doNondominatedSorting}
\alias{doNondominatedSorting}
\title{Fast non-dominated sorting algorithm.}
\usage{
doNondominatedSorting(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric matrix of points. Each row contains one objective vector.}
}
\value{
[\code{list}]
  List with the following components
  \describe{
    \item{ranks}{Integer vector of ranks of length \code{nrow(x)}. The higher
    the rank, the higher the domination front the corresponding points is
    located on.}
    \item{dom.counter}{Integer vector of length \code{nrow(x)}. The i-th element
    is the domination number of the i-th point.}
  }
}
\description{
Pure R implementation of the fast non-dominated sorting algorithm
proposed by Deb.
}
\references{
Deb, K., Pratap, A., and Agarwal, S. A Fast and Elitist Multiobjective Genetic
Algorithm: NSGA-II. IEEE Transactions on Evolutionary Computation, 6 (8) (2002),
182-197.
}

