% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecr.R
\name{ecr}
\alias{ecr}
\title{Working horse of the ecr package.}
\usage{
ecr(objective.fun, control)
}
\arguments{
\item{objective.fun}{[\code{smoof_function}]\cr
Single objective target function of type \code{smoof_function}.}

\item{control}{[\code{ecr.control}]\cr
Control object.}
}
\value{
[\code{ecrResult}]
  Object of type \code{ecr_result} containing a list:
  \itemize{
   \item{objective.fun \code{smoof_function}}{Objective function.}
   \item{control \code{ect_control}}{Control object.}
   \item{best.param \code{numeric}}{Best parameter combination.}
   \item{best.value \code{numeric(1)}}{Best reached value.}
   \item{opt.path \code{optPath}}{Optimization path.}
   \item{population.storage \code{list}}{List of populations.}
   \item{message \code{character(1)}}{Message explaining the reason for termination.}
  }
}
\description{
Takes a function and searches for a global optimum with an evolutionary approach.
}
\examples{
library(smoof)
library(ParamHelpers)
# We want to find the minimum of the function f(x) = x sin(2x) on the intervall
# [0, 2pi]. The optimal value is about -5.5 for x = 5.54.
# First we wrap the function with the smoof package:
obj.fn = makeSingleObjectiveFunction(
  name = "My obj. function",
  fn = function(x) x * sin(2 * x),
  par.set = makeParamSet(makeNumericParam("x", lower = 0, upper = 2 * pi))
)

# We want to solve this with a (10 + 10) evolutionary strategy based on
# the floating point representation of the input vectors with the default
# operators: intermediate recombinator and Gauss mutation
ctrl = ecr.control(
  population.size = 10L,
  offspring.size = 10L,
  survival.strategy = "plus",
  n.params = 1L,
  representation = "float",
  stopping.conditions = setupStoppingConditions(max.iter = 100L)
)

res = ecr(obj.fn, control = ctrl)
print(res)
}
\seealso{
\code{\link{ecr.control}}
}

