% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecr.R
\name{ecr}
\alias{ecr}
\title{Working horse of the ecr package.}
\usage{
ecr(objective.fun, control)
}
\arguments{
\item{objective.fun}{[\code{smoof_function}]\cr
Single objective target function of type \code{smoof_function}.}

\item{control}{[\code{ecr.control}]\cr
  Control object.
return [\code{ecrResult}]
  Object of type \code{ecr_result} containing a list:
  \itemize{
   \item{objective.fun \code{smoof_function}}{Objective function.}
   \item{control \code{ect_control}}{Control object.}
   \item{best.param \code{numeric}}{Best parameter combination.}
   \item{best.value \code{numeric(1)}}{Best reached value.}
   \item{opt.path \code{optPath}}{Optimization path.}
   \item{population.storage \code{list}}{List of populations.}
   \item{message \code{character(1)}}{Message explaining the reason for termination.}
  }}
}
\description{
Takes a function and searches for a global optimum with an evolutionary approach.
}

