% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ecr}
\alias{ecr}
\title{Working horse of the ecr package.}
\usage{
ecr(objective.fun, control, global.optimum = NA)
}
\arguments{
\item{objective.fun}{[\code{function}]\cr
Target function.}

\item{control}{[\code{ecr.control}]\cr
Control object.}

\item{global.optimum}{[\code{numeric}]\cr
  Parameter combination of the global optimum of the fun. This parameter is optional.
return [\code{ecrResult}]
  Object of type \code{ecrResult} containing a list:
  \itemize{
   \item{best.param \code{numeric}}{Best parameter combination.}
   \item{best.value \code{numeric(1)}}{Best reached value.}
   \item{trace \code{ecrTrace}}{Optimization path.}
  }}
}
\description{
Takes a function and searches for global optimum with an evolutionary approach.
}

