% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeECRResult.R
\name{ECRMultiObjectiveResult}
\alias{ECRMultiObjectiveResult}
\title{Multi objective result object.}
\description{
Object returned by \code{\link{doTheEvolution}} in case of the objective function
being multi-objective.
}
\details{
It contains ...
}

