% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutator.gauss.R
\docType{data}
\name{gaussMutator}
\alias{gaussMutator}
\title{Generator of the Gauss mutation operator.}
\format{\preformatted{function (setOfIndividuals, control = list(mutator.gauss.prob = 1, mutator.gauss.sd = 0.05))  
 - attr(*, "name")= chr "Gauss mutator"
 - attr(*, "description")= chr "Adds gaussian noise to each gene"
 - attr(*, "supported")= chr "float"
 - attr(*, "class")= chr [1:2] "ecr_operator" "ecr_mutator"
 - attr(*, "defaults")=List of 2
  ..$ mutator.gauss.prob: num 1
  ..$ mutator.gauss.sd  : num 0.05
 - attr(*, "checkFunction")=function (operator.control)  
}}
\usage{
gaussMutator
}
\arguments{
\item{setOfIndividuals}{[\code{setOfIndividuals}]\cr
Set of individuals.}

\item{control}{[\code{list}]\cr
List containing evolutionary operators for fine-tuning the mutation operator.
The \code{gaussMutator} function expects the two values:
\itemize{
  \item{mutator.gauss.prob [\code{numeric(1)}]}{Probability of mutation for the gauss mutation operator.}
  \item{mutator.gauss.sd [\code{numeric(1)}}{Standard deviance of the Gauss mutation, i. e., the mutation strength.}
}}
}
\value{
[\code{setOfIndividuals}]
  Set of individuals.
}
\description{
Default Gauss mutation operator known from Evolutionary Algorithms.
}
\keyword{datasets}

