% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getBestIndividual}
\alias{getBestIndividual}
\title{Extracts the best individual of a set of individuals.}
\usage{
getBestIndividual(setOfIndividuals)
}
\arguments{
\item{setOfIndividuals}{[\code{setOfIndividuals}]\cr
Set of individuals (population).}
}
\value{
[\code{essoIndividual}]
 Individual with best, i. e., lowest fitness value.
}
\description{
Extracts the best individual of a set of individuals.
}

