% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestIndividual.R
\name{getBestIndividual}
\alias{getBestIndividual}
\title{Extracts the best individual of a set of individuals.}
\usage{
getBestIndividual(population)
}
\arguments{
\item{population}{[\code{setOfIndividuals}]\cr
Population.}
}
\value{
[list]
 List with the following entries
 \describe{
   \item{individual}{Genotype of the best individual.}
   \item{fitness}{Fitness of the individual.}
 }
}

