% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getBestIndividual.R
\name{getBestIndividual}
\alias{getBestIndividual}
\title{Extracts the best individual of a set of individuals.}
\usage{
getBestIndividual(population)
}
\arguments{
\item{population}{[\code{setOfIndividuals}]\cr
Population.}
}
\value{
[\code{essoIndividual}]
 Individual with best, i. e., lowest fitness value.
}
\description{
Extracts the best individual of a set of individuals.
}

