% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeECRResult.R
\name{getEvaluations}
\alias{getEvaluations}
\title{Determine number of function evaluations needed.}
\usage{
getEvaluations(result)
}
\arguments{
\item{result}{[\code{ecr_result}]\cr
\pkg{ecr} result object.}
}
\value{
[integer(1)]
}
\description{
Determine number of function evaluations needed.
}

