% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{getEvaluations}
\alias{getEvaluations}
\title{Get number of function evaluations.}
\usage{
getEvaluations(result)
}
\arguments{
\item{result}{[\code{ecr_result}]\cr
\pkg{ecr} result object.}
}
\value{
[integer(1)]
}
\description{
Determine the number of function evaluations needed by the EA.
}

