\name{getOperatorName}
\alias{getOperatorName}
\title{Determine the name of a given operator.}
\usage{
getOperatorName(operator)
}
\arguments{
  \item{operator}{[\code{esoo_operator}]\cr Operator
  object.}
}
\value{
[\code{character(1)}] Name of the operator.
}
\description{
Determine the name of a given operator.
}

