% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{getOperatorName}
\alias{getOperatorName}
\title{Determine the name of a given operator.}
\usage{
getOperatorName(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{character(1)}]
  Name of the operator.
}
\description{
Determine the name of a given operator.
}

