% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getOperatorsTags.R
\name{getSupportedRepresentations}
\alias{getSupportedRepresentations}
\title{Returns the character vector of tags which describe a specific operator.}
\usage{
getSupportedRepresentations(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{character}]
  Vector of representation types.
}
\description{
Returns the character vector of tags which describe a specific operator.
}

