% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{intermediateRecombinator}
\alias{intermediateRecombinator}
\title{Generator of the indermediate recombination operator.}
\format{\preformatted{function (setOfIndividuals, control = list())  
 - attr(*, "name")= chr "Intermediate recombinator"
 - attr(*, "description")= chr "No description"
 - attr(*, "supported")= chr "float"
 - attr(*, "n.parents")= int 10
 - attr(*, "defaults")= list()
 - attr(*, "class")= chr [1:2] "ecr_operator" "ecr_recombinator"
 - attr(*, "checkFunction")=function (operator.control)  
}}
\usage{
intermediateRecombinator
}
\arguments{
\item{setOfIndividuals}{[\code{setOfIndividuals}]\cr
Set of individuals.}

\item{control}{[\code{list}]\cr
Empty list. Intermediate recombinator has no paramerters.}
}
\value{
[\code{setOfIndividuals}]
  Recombined offspring.
}
\description{
Generator of the indermediate recombination operator.
}
\keyword{datasets}

