% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.random.binary.R
\name{makeBinaryGenerator}
\alias{makeBinaryGenerator}
\title{Generates a generator object for the initial population.}
\usage{
makeBinaryGenerator()
}
\value{
[\code{ecr_generator}]
}
\description{
The generated operator samples uniformally distributed points in the design
space of the target function taking care not to violate box constraints.
}

