% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutator.bitflip.R
\name{makeBitFlipMutator}
\alias{makeBitFlipMutator}
\title{Generator of a simple bitplip mutation operator.}
\usage{
makeBitFlipMutator(mutator.flip.prob = 0.1)
}
\arguments{
\item{mutator.flip.prob}{[\code{numeric(1)}]\cr
Probability to flip a single bit. Default is \code{0.1}.}
}
\value{
[\code{ecr_mutator}]
}
\description{
This operator works only on binary representation and flips each bit
  with a given probability.
}

