% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeConsoleMonitor.R
\name{makeConsoleMonitor}
\alias{makeConsoleMonitor}
\title{Simple monitoring function.}
\usage{
makeConsoleMonitor(show.info.stepsize = 5L)
}
\arguments{
\item{show.info.stepsize}{[\code{integer(1)}]\cr
Adjust the stepsize of iterations with informative messages.}
}
\value{
[\code{ecr_monitor}]
}
\description{
This is the default monitoring function used by ecr. It simply outputs
the iteration as well as minimal, mean and maximal target values from the current
population.
}

