% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeConsoleMonitor.R
\name{makeConsoleMonitor}
\alias{makeConsoleMonitor}
\title{Simple monitoring function.}
\usage{
makeConsoleMonitor()
}
\description{
This is the default monitoring function used by ecr. It simply outputs
the iteration as well as minimal, mean and maximal target values from the current
population.
}

