% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recombinator.crossover.R
\name{makeCrossoverRecombinator}
\alias{makeCrossoverRecombinator}
\title{Generator of the crossover recombination operator.}
\usage{
makeCrossoverRecombinator()
}
\value{
[\code{ecr_recombinator}]
}
\description{
Generator of the crossover recombination operator.
}

