% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinator.crossover.R
\name{makeCrossoverRecombinator}
\alias{makeCrossoverRecombinator}
\title{Generator of the One-point crossover recombination operator.}
\usage{
makeCrossoverRecombinator(p = 1)
}
\arguments{
\item{p}{[\code{numeric(1)}]\cr
Cross over probability to form an offspring. Default is \code{1}.}
}
\value{
[\code{ecr_recombinator}]
}
\seealso{
Other recombinators: \code{\link{makeIntermediateRecombinator}},
  \code{\link{makeNullRecombinator}},
  \code{\link{makeOXRecombinator}},
  \code{\link{makePMXRecombinator}}
}

