% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutator.gauss.R
\name{makeGaussMutator}
\alias{makeGaussMutator}
\title{Generator of the Gauss mutation operator.}
\usage{
makeGaussMutator(mutator.gauss.prob = 1L, mutator.gauss.sd = 0.05)
}
\arguments{
\item{mutator.gauss.prob}{[\code{numeric(1)}]\cr
Probability of mutation for the gauss mutation operator.}

\item{mutator.gauss.sd}{[\code{numeric(1)}\cr
Standard deviance of the Gauss mutation, i. e., the mutation strength.}
}
\value{
[\code{ecr_mutator}]
}
\description{
Default Gauss mutation operator known from Evolutionary Algorithms.
}

