% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selector.greedy.R
\name{makeGreedySelector}
\alias{makeGreedySelector}
\title{Simple \dQuote{greedy} selector.}
\usage{
makeGreedySelector()
}
\value{
[\code{setOfIndividuals}]
}
\description{
Sorts the individuals according to their fitness value in increasing order
  and selects the best ones.
}

