% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stoppingCondition.max.time.R
\name{makeMaximumTimeStoppingCondition}
\alias{makeMaximumTimeStoppingCondition}
\title{Stopping condition: time limit.}
\usage{
makeMaximumTimeStoppingCondition(max.time = NULL)
}
\arguments{
\item{max.time}{[\code{integer(1)}]\cr
Time budget in seconds. Default ist \code{Inf}.}
}
\value{
[\code{function}]
}
\description{
Stopping condition: time limit.
}

