% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMutator.R
\name{makeMutator}
\alias{makeMutator}
\title{Construct a mutation operator.}
\usage{
makeMutator(mutator, name, description,
  supported = getAvailableRepresentations(), defaults = list(),
  checker = function(operator.control) TRUE)
}
\arguments{
\item{mutator}{[\code{function}]\cr
Actual mutation operator.}

\item{name}{[\code{character(1)}]\cr
Name of the mutator.}

\item{description}{[\code{character(1)}]\cr
Short description of how the mutator works.}

\item{supported}{[\code{character}]\cr
Vector of strings/names of supported parameter representations. For example
'permutation', 'float', 'binary'.}

\item{defaults}{[\code{list}]\cr
List of default strategy parameter values.}

\item{checker}{[\code{function}]\cr
Check object, which performs a sanity check in mutator strategy parameters
passed to the control object.}
}
\value{
[\code{ecr_mutator}]
  Mutator object.
}
\description{
Helper function which constructs a mutator, i. e., a mutation operator.
}

