% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeNullMonitor.R
\name{makeNullMonitor}
\alias{makeNullMonitor}
\title{Simple no output function.}
\usage{
makeNullMonitor()
}
\value{
[\code{ecr_monitor}]
}
\description{
No monitoring at all. Using this monitor practically deactivates monitoring.
}

