% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombinator.null.R
\name{makeNullRecombinator}
\alias{makeNullRecombinator}
\title{The nullRecombinator does not perform any recombination.}
\usage{
makeNullRecombinator()
}
\value{
[\code{ecr_recombinator}]
}
\description{
The nullRecombinator does not perform any recombination.
}

