% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makePermutationGenerator}
\alias{makePermutationGenerator}
\title{Generates a generator object for the initial population.}
\usage{
makePermutationGenerator()
}
\value{
[\code{ecr_opeator}]
}
\description{
Generates a generator object for the initial population.
}

