% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinator.pmx.R
\name{makePMXRecombinator}
\alias{makePMXRecombinator}
\title{Generator of the Partially-Mapped-Crossover (PMX) recombination operator.}
\usage{
makePMXRecombinator()
}
\value{
[\code{ecr_recombinator}]
}
\description{
This recombination operator is specifically designed for permutations.
  The operators chooses two cut-points at random and generates two offspring
  as follows: a) copy the subsequence of one parent and b) fill the remaining
  positions while preserving the order and position of as many genes as possible.
}
\seealso{
Other recombinators: \code{\link{makeCrossoverRecombinator}},
  \code{\link{makeIntermediateRecombinator}},
  \code{\link{makeNullRecombinator}},
  \code{\link{makeOXRecombinator}}
}

