% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makePopulation.R
\name{makePopulation}
\alias{makePopulation}
\title{Generate population from individuals}
\usage{
makePopulation(individuals, fitness = NULL)
}
\arguments{
\item{individuals}{[\code{list}]\cr
List of individuals.}

\item{fitness}{[\code{matrix}]\cr
Matrix with one row of fitness values for the individuals.
Default is \code{NULL}.}
}
\value{
[\code{setOfIndividuals}]
}
\description{
Helper function for wrapping individuals in a 'setOfIndividuals'.
}

