% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePopulation.R
\name{makePopulation}
\alias{makePopulation}
\title{Generate a population.}
\usage{
makePopulation(individuals, fitness = NULL)
}
\arguments{
\item{individuals}{[\code{matrix}]\cr
List of individuals.}

\item{fitness}{[\code{matrix}]\cr
Matrix with one row of fitness values for the individuals.
Default is \code{NULL}.}
}
\value{
[\code{setOfIndividuals}]
}
\description{
Wrap individuals in a \code{setOfIndividuals}. This function is of particular
interest when one wants to write population generators for non-standard
representations.
}

