% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.scramble.R
\name{makeScrambleMutator}
\alias{makeScrambleMutator}
\title{Generator for the Scramble mutation operator.}
\usage{
makeScrambleMutator()
}
\value{
[\code{ecr_mutator}]
}
\description{
The Scramble mutation operator selects two positions within the chromosome at
  random and randomly intermixes the subsequence between these positions.
}
\seealso{
Other mutators: \code{\link{makeBitFlipMutator}},
  \code{\link{makeGaussMutator}},
  \code{\link{makeInsertionMutator}},
  \code{\link{makeSwapMutator}},
  \code{\link{makeUniformMutator}}
}

