% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeSelector.R
\name{makeSelector}
\alias{makeSelector}
\title{Construct a selection operator.}
\usage{
makeSelector(selector, name, description,
  supported = getAvailableRepresentations(), supported.objectives)
}
\arguments{
\item{selector}{[\code{function}]\cr
Actual selection operator.}

\item{name}{[\code{character(1)}]\cr
Name of the selector.}

\item{description}{[\code{character(1)}]\cr
Short description of how the selector works.}

\item{supported}{[\code{character}]\cr
Vector of strings/names of supported parameter representations. For example
'permutation', 'float', 'binary'.}

\item{supported.objectives}{[\code{character}]\cr
At least one of \dQuote{single-objective} or \dQuote{multi-objective}.}
}
\value{
[\code{ecr_selector}]
  selector object.
}
\description{
Helper function which defines a selector method, i. e., an operator which
  takes the population and returns a part of it for mating.
}

