% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uniformPopulationGenerator.R
\name{makeUniformGenerator}
\alias{makeUniformGenerator}
\title{Generates a generator object for the initial population.}
\usage{
makeUniformGenerator()
}
\value{
[\code{ecr_opeator}]
}
\description{
Generates a generator object for the initial population.
}

