% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeUniformGenerator}
\alias{makeUniformGenerator}
\title{Generates a generator object for the initial population.}
\usage{
makeUniformGenerator()
}
\value{
[\code{ecr_opeator}]
}
\description{
Generates a generator object for the initial population.
}

