% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.uniform.float.R
\name{makeUniformGenerator}
\alias{makeUniformGenerator}
\title{Generates the uniform generator object for the initial population.}
\usage{
makeUniformGenerator()
}
\value{
[\code{ecr_generator}]
}
\description{
The returned population contains individuals which are uniformly distributed
within the bounds specified by the paramter set of the \pkg{smoof} objective
function passed to the \code{\link{doTheEvolution}} function.
}

