% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePopulations.R
\name{mergePopulations}
\alias{mergePopulations}
\title{Merge populations.}
\usage{
mergePopulations(...)
}
\arguments{
\item{...}{[\code{list}]\cr
List of objects of type \code{setOfIndividuals}.}
}
\value{
[\code{setOfIndividuals}]
}
\description{
Make one population out of multiple.
}

