\name{nullRecombinator}
\alias{nullRecombinator}
\title{The nullRecombinator does not perform any recombination.}
\usage{
nullRecombinator(setOfIndividuals, control = list())
}
\arguments{
  \item{setOfIndividuals}{[\code{setOfIndividuals}]\cr Set
  of individuals.}

  \item{control}{[\code{list}]\cr Empty list. Intermediate
  recombinator has no paramerters.}
}
\value{
[\code{setOfIndividuals}] Recombined offspring.
}
\description{
The nullRecombinator does not perform any recombination.
}

