% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nullRecombinator.R
\docType{data}
\name{nullRecombinator}
\alias{nullRecombinator}
\title{The nullRecombinator does not perform any recombination.}
\format{\preformatted{function (setOfIndividuals, control = list())  
 - attr(*, "name")= chr "null recombinator"
 - attr(*, "description")= chr "No description"
 - attr(*, "supported")= chr [1:4] "permutation" "binary" "float" "integer"
 - attr(*, "n.parents")= int 10
 - attr(*, "defaults")= list()
 - attr(*, "class")= chr [1:2] "ecr_operator" "ecr_recombinator"
 - attr(*, "checkFunction")=function (operator.control)  
}}
\usage{
nullRecombinator
}
\arguments{
\item{setOfIndividuals}{[\code{setOfIndividuals}]\cr
Set of individuals.}

\item{control}{[\code{list}]\cr
Empty list. Intermediate recombinator has no paramerters.}
}
\value{
[\code{setOfIndividuals}]
  Recombined offspring.
}
\description{
The nullRecombinator does not perform any recombination.
}
\keyword{datasets}

