% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupECRControl.R
\name{print.ecr_control}
\alias{print.ecr_control}
\title{Print ecr control object.}
\usage{
\method{print}{ecr_control}(x, ...)
}
\arguments{
\item{x}{[\code{ecr_control}]\cr
Control object.}

\item{...}{[any]\cr
Not used.}
}
\description{
Print ecr control object.
}

