% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupEvolutionaryOperators.R
\name{setupEvolutionaryOperators}
\alias{setupEvolutionaryOperators}
\title{Define the toolbix of all the evolutionary operators you wish to operate on
your problem.}
\usage{
setupEvolutionaryOperators(control,
  parent.selector = getDefaultEvolutionaryOperators(control$representation,
  "parent.selector"),
  survival.selector = getDefaultEvolutionaryOperators(control$representation,
  "survival.selector"),
  generator = getDefaultEvolutionaryOperators(control$representation,
  "generator"),
  mutator = getDefaultEvolutionaryOperators(control$representation,
  "mutator"), mutationStrategyAdaptor = function(operator.control, opt.path) {
      return(operator.control) },
  recombinator = getDefaultEvolutionaryOperators(control$representation,
  "recombinator"), mutator.control = list(), recombinator.control = list())
}
\arguments{
\item{control}{[\code{ecr_control}]\cr
ECR control object generated via \code{\link{setupECRControl}}.}

\item{parent.selector}{[\code{ecr_selector}]\cr
Selection operator which implements a procedure to copy individuals from a
given population to the mating pool, i. e., allow them to become parents.}

\item{survival.selector}{[\code{ecr_selector}]\cr
Selection operator which implements a procedurce to extract individuals from
a given set, which should survive and set up the next generation.}

\item{generator}{[\code{ecr_generator}]\cr
Generator operator of type \code{ecr_generator} for the generation of the initial
population.}

\item{mutator}{[\code{ecr_mutator}]\cr
Mutation operator of type \code{ecr_mutator}.}

\item{mutationStrategyAdaptor}{[\code{function}]\cr
This is an experimental parameter. Hence, you should be careful when using it.
Serves to offer the possibility to adapt parameters of the mutation algorithm
(e. g. mutation stepsize \eqn{\sigma} for Gaussian mutation) in each iteration.
The function needs to expect the parameters \dQuote{operator.control} and
\dQuote{opt.path}, the last being of type \code{\link[ParamHelpers]{OptPath}} and
must return the modified \dQuote{operator.control} object. The default does
nothing.}

\item{recombinator}{[\code{ecr_recombinator}]\cr
Recombination operator of type \code{ecr_recombinator}.}

\item{mutator.control}{[\code{list}]\cr
List of evolutionary parameters for the corresponding mutation operator. See the
help pages for the mutation operators for the needed values.}

\item{recombinator.control}{[\code{list}]\cr
List of evolutionary parameters for the corresponding recombination operator. See the
help pages for the recombination operators for the needed values.}
}
\value{
[\code{ecr_control}] Modified control object.
}
\description{
Define the toolbix of all the evolutionary operators you wish to operate on
your problem.
}
\note{
Keep in mind, that all of the provided operators need to be compatible
with the \dQuote{representation} stored in the \code{control} object.
}

