% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutator.swap.R
\docType{data}
\name{swapMutator}
\alias{swapMutator}
\title{Swap mutation operator.}
\format{\preformatted{function (setOfIndividuals, control = list())  
 - attr(*, "name")= chr "Swap mutator"
 - attr(*, "description")= chr "Swaps two alleles"
 - attr(*, "supported")= chr "permutation"
 - attr(*, "class")= chr [1:2] "ecr_operator" "ecr_mutator"
 - attr(*, "defaults")= list()
 - attr(*, "checkFunction")=function (operator.control)  
}}
\usage{
swapMutator
}
\arguments{
\item{setOfIndividuals}{[\code{setOfIndividuals}]\cr
Set of individuals.}

\item{control}{[\code{list}]\cr
Not used.}
}
\value{
[\code{setOfIndividuals}]
  Set of individuals.
}
\description{
Swap mutation operator.
}
\keyword{datasets}

